Add errors to many actions

+++ src/realworld/actions/comment/delete.cr
@@ -1,4 +1,5 @@
 require "../base"
+require "../../errors"
 require "../../models/article"
 require "../../models/comment"
 require "../../services/repo"
@@ -10,23 +11,32 @@ module Realworld::Actions::Comment
     
     def call(env)
       user = env.get("auth").as(User)
-      article = Repo.get_by(Article, slug: env.params.url["slug"])
-      if article
-        if id = env.params.url["slug"].to_i64?
-          comment = Repo.get(Comment, id)
-          if comment && article.id == comment.article_id && user.id == comment.user_id
-            changeset = Repo.delete(comment)
 
-            # TODO: return success
-          else
-            # TODO: return error
+      slug = env.params.url["slug"]
+      id   = env.params.url["id"].to_i64?
+      
+      raise Realworld::NotFoundException.new(env) if !id
+      
+      article = Repo.get_by(Article, slug: slug)
+      raise Realworld::NotFoundException.new(env) if !article
+      
+      comment = Repo.get(Comment, id)
+      raise Realworld::NotFoundException.new(env) if !comment
+
+      if article.id == comment.article_id && user.id == comment.user_id
+        changeset = Repo.delete(comment)
+        if !changeset.valid?
+          errors = {} of String => Array(String)
+          changeset.errors.reduce(errors) do |memo, error|
+            memo[error[:field]] = memo[error[:field]]? || [] of String
+            memo[error[:field]] << error[:message]
+            memo
           end
-        else
-          # TODO: return error
+          raise Realworld::UnprocessableEntityException.new(env, errors)
         end
       else
-        # TODO: return error
-      end  
+        raise Realworld::ForbiddenException.new(env)
+      end
     end
   end
 end
